requirements.txt
aiogram==3.14.0
aiosqlite==0.22.1
google-generativeai==0.8.3
